TEST-3 <Compplete Bubble Sort Funtion>

+++ src/Main.java
@@ -12,12 +12,12 @@ public class Main {
     public static void sayHello(){
         System.out.println("Hello Con Cho");
     }
-    public static void swap(int a, int b){
-        int temp = b;
+    public static void swap(float a, float b){
+        float temp = b;
         b = a;
         a = temp;
     }
-    public static void bubbleSort(int[] a){
+    public static void bubbleSort(float[] a){
         for (int i = 0; i < a.length - 1; i++)
             for(int j = 0; j< a.length - 1; j ++ ){
                 if (a[i] > a[j]){
